[ververica] extract common job setup into a single utility class
