+++ troubleshooting/introduction/src/solution/java/com/ververica/flink/training/solutions/TroubledStreamingJobSolution2.java
@@ -28,7 +28,8 @@ import com.ververica.flink.training.common.WindowedMeasurements;
 import java.io.IOException;
 import java.util.concurrent.TimeUnit;
 
-import static com.ververica.flink.training.exercises.TroubledStreamingJobUtils.createConfiguredEnvironment;
+import static com.ververica.flink.training.common.EnvironmentUtils.createConfiguredEnvironment;
+import static com.ververica.flink.training.common.EnvironmentUtils.isLocal;
 
 /**
  * Troubled streaming job evolved on top of {@link TroubledStreamingJobSolution1} by fixing the
@@ -44,9 +45,7 @@ public class TroubledStreamingJobSolution2 {
 	public static void main(String[] args) throws Exception {
 		ParameterTool parameters = ParameterTool.fromArgs(args);
 
-		final boolean local = parameters.getBoolean("local", false);
-
-		StreamExecutionEnvironment env = createConfiguredEnvironment(parameters, local);
+		StreamExecutionEnvironment env = createConfiguredEnvironment(parameters);
 
 		//Time Characteristics
 		env.setStreamTimeCharacteristic(TimeCharacteristic.EventTime);
@@ -81,7 +80,7 @@ public class TroubledStreamingJobSolution2 {
 				.name("WindowedAggregationPerLocation")
 				.uid("WindowedAggregationPerLocation");
 
-		if (local) {
+		if (isLocal(parameters)) {
 			aggregatedPerLocation.print()
 					.name("NormalOutput")
 					.uid("NormalOutput")
