+++ troubleshooting/introduction/src/solution/java/com/ververica/flink/training/solutions/TroubledStreamingJobSolution42.java
@@ -29,7 +29,8 @@ import com.ververica.flink.training.common.WindowedMeasurements;
 import java.io.IOException;
 import java.util.concurrent.TimeUnit;
 
-import static com.ververica.flink.training.exercises.TroubledStreamingJobUtils.createConfiguredEnvironment;
+import static com.ververica.flink.training.common.EnvironmentUtils.createConfiguredEnvironment;
+import static com.ververica.flink.training.common.EnvironmentUtils.isLocal;
 
 /**
  * Troubled streaming job evolved on top of {@link TroubledStreamingJobSolution41} by optimising
@@ -45,9 +46,7 @@ public class TroubledStreamingJobSolution42 {
 	public static void main(String[] args) throws Exception {
 		ParameterTool parameters = ParameterTool.fromArgs(args);
 
-		final boolean local = parameters.getBoolean("local", false);
-
-		StreamExecutionEnvironment env = createConfiguredEnvironment(parameters, local);
+		StreamExecutionEnvironment env = createConfiguredEnvironment(parameters);
 
 		//Time Characteristics
 		env.setStreamTimeCharacteristic(TimeCharacteristic.EventTime);
@@ -84,7 +83,7 @@ public class TroubledStreamingJobSolution42 {
 				.name("WindowedAggregationPerLocation")
 				.uid("WindowedAggregationPerLocation");
 
-		if (local) {
+		if (isLocal(parameters)) {
 			aggregatedPerLocation.print()
 					.name("NormalOutput")
 					.uid("NormalOutput")
