+++ troubleshooting/object-reuse/src/main/java/com/ververica/flink/training/exercises/ObjectReuseJob.java
@@ -27,7 +27,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
-import static com.ververica.flink.training.exercises.ObjectReuseJobUtils.createConfiguredEnvironment;
+import static com.ververica.flink.training.common.EnvironmentUtils.createConfiguredEnvironment;
 
 /**
  * Object reuse exercise.
@@ -42,9 +42,7 @@ public class ObjectReuseJob {
 	public static void main(String[] args) throws Exception {
 		ParameterTool parameters = ParameterTool.fromArgs(args);
 
-		final boolean local = parameters.getBoolean("local", false);
-
-		StreamExecutionEnvironment env = createConfiguredEnvironment(parameters, local);
+		StreamExecutionEnvironment env = createConfiguredEnvironment(parameters);
 
 		final boolean objectReuse = parameters.getBoolean("objectReuse", false);
 		if (objectReuse) {
