+++ troubleshooting/object-reuse/src/solution/java/com/ververica/flink/training/solutions/ObjectReuseJobSolution1.java
@@ -29,7 +29,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
-import static com.ververica.flink.training.exercises.ObjectReuseJobUtils.createConfiguredEnvironment;
+import static com.ververica.flink.training.common.EnvironmentUtils.createConfiguredEnvironment;
 
 /**
  * Object reuse solution fixing incorrect object reuse.
@@ -44,9 +44,7 @@ public class ObjectReuseJobSolution1 {
 	public static void main(String[] args) throws Exception {
 		ParameterTool parameters = ParameterTool.fromArgs(args);
 
-		final boolean local = parameters.getBoolean("local", false);
-
-		StreamExecutionEnvironment env = createConfiguredEnvironment(parameters, local);
+		StreamExecutionEnvironment env = createConfiguredEnvironment(parameters);
 
 		final boolean objectReuse = parameters.getBoolean("objectReuse", false);
 		if (objectReuse) {
